# packages/shared/openapi/agent-mission-control.openapi.yaml
openapi: 3.0.3
info:
  title: Agent Mission Control API
  version: 0.2.0
  description: >-
    Agentic AI branding/marketing agency execution API with local and production
    runtime modes. Production authentication uses Supabase bearer tokens plus
    server-side tenant/project memberships. External publication and paid-media
    execution remain fail-closed until concrete provider adapters are installed.
security:
  - BearerAuth: []
paths:
...
  /approvals:
    get:
      summary: List pending approvals for the authenticated tenant
      parameters:
        - in: query
          name: tenant_id
          required: false
          deprecated: true
          description: Optional compatibility assertion; if supplied it must match the authenticated tenant.
          schema: { type: string }
      responses:
        '200':
          description: Pending approval rows for the authenticated tenant
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ApprovalRequest' }
        '401': { description: Authentication required or token invalid }
        '403': { description: Supplied tenant_id does not match authenticated tenant }
        '503': { description: Authentication provider or runtime configuration unavailable }
  /approvals/{approval_id}/decide:
    post:
      summary: Resolve a pending approval exactly once
      parameters:
        - in: path
          name: approval_id
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecision'
      responses:
        '200':
          description: Immutable approval decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalDecisionResponse'
        '400': { description: Invalid decision }
        '401': { description: Authentication required or token invalid }
        '403': { description: Approval belongs to another authorization scope }
        '404': { description: Approval not found }
        '409': { description: Approval is terminal, run state changed, or idempotency conflict }
        '503': { description: Authentication provider or runtime configuration unavailable }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Supabase access token
  parameters:
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 1 }
  schemas:
...
    ApprovalRequest:
      type: object
      required: [approval_id, run_id, tenant_id, project_id, reason, status, created_at]
      properties:
        approval_id: { type: string, format: uuid }
        run_id: { type: string }
        tenant_id: { type: string }
        project_id: { type: string }
        reason: { type: string }
        confidence: { type: number, nullable: true }
        status: { type: string, enum: [pending, resolved, stale] }
        reviewer: { type: string, nullable: true }
        decision: { type: string, enum: [approve, reject], nullable: true }
        subject_type: { type: string, nullable: true }
        subject_ref: { type: string, nullable: true }
        subject_version_ref: { type: string, nullable: true }
        subject_hash: { type: string, nullable: true }
        authority_ref: { type: string, nullable: true }
        policy_version: { type: string, nullable: true }
        stale_reason: { type: string, nullable: true }
        staled_at: { type: string, nullable: true }
        created_at: { type: string }
        decided_at: { type: string, nullable: true }
    ApprovalDecision:
      type: object
      required: [decision]
      properties:
        decision: { type: string, enum: [approve, reject] }
    ApprovalDecisionResponse:
      type: object
      required: [approval_id, status, decision, reviewer, decided_at]
      properties:
        approval_id: { type: string, format: uuid }
        status: { type: string, enum: [resolved] }
        decision: { type: string, enum: [approve, reject] }
        reviewer: { type: string }
        decided_at: { type: string }
...